Merged changeset r7936:7943 from jcorgan/bc-merge into release branch.
This merges in Michael Dicken's build system updates for ticket:186 into
the release 3.1 branch.  It effectively combines changesets r7617, r7631, 
r7632, r7769, and r7888 from the trunk.

Merge conflicts resulted from portions of changesets that applied to
files that are not on the release branch, and were manually resolved.

File config/pkg.m4 was copied over from the trunk.


git-svn-id: http://nyquist.gnuradio.org/svn/gnuradio/branches/releases/3.1@7945 221aa14e-8319-0410-a670-987f0aec2ac5